# Replace bytes.NewReader(stringToBytes(s)) with strings.NewReader(s)